Reforging main function

+++ script/stereo.py
@@ -21,7 +21,7 @@ def parse_cmdline():
                         help='Set left port Name. Port Name has to be between { A, B, C, D}')
     parser.add_argument('-d', '--window-name', type=str, nargs=None, required=True,
                         help='Set window header name.')
-    parser.add_argument('-f', '--framerate', type=float, nargs=None, required=True,
+    parser.add_argument('-f', '--framerate', type=int, nargs=None, required=True,
                         help='Capture framerate not specified Default 20Hz will be used.')
     parser.add_argument('-w', '--width', type=int, nargs=None, required=True,
                         help='Set capture width with argument -w.')
